custom_data: remove procfile and fix processes name in registerUnit
We don't need to provide the procfile, we used to provide it for older
versions of tsuru, but we're not interested in supporting old versions
of tsuru.

I believe that this is related to tsuru/tsuru#1302, but I'm not closing
the issue yet, as I cannot prove that this commit fixes the issue.