all: adds image deploy support for kubernetes
This commit adds support for sidecar image deployment. When SOURCE_IMAGE
is provided, the sidecar will inspect the given image and also
a potential Procfile and tsuru.yaml on the primary container. This is
sent to stdout so tsuru API can parse and process them.

The primary container is commited and it's image is pushed.